requirements.txt
fastapi
uvicorn
httpx
pydantic
pydantic-settings
python-dotenv
slowapi
sqlalchemy[asyncio]
aiosqlite
asyncpg
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pytest
pytest-asyncio
alembic
openai
aiodns
fpdf2
